requirements.txt
numpy==1.23.1
opencv_python==4.6.0.66
pdf2image==1.17.0
Pillow==9.2.0
Pillow==11.0.0
pytesseract==0.3.13
